% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_2D}
\alias{approx_2D}
\title{Perform 2D linear approximation}
\usage{
approx_2D(x, y, xout)
}
\arguments{
\item{x}{Vector, giving the coordinates of the points to be interpolated.}

\item{y}{Matrix, giving the points to be interpolated.}

\item{xout}{Vector, values specifying where interpolation is to take place.}
}
\value{
Matrix, interpolated data points.
}
\description{
Return a matrix of points which linearly interpolate given data points.
}
