% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{covariance_ll}
\alias{covariance_ll}
\title{Perform an estimation of the covariance with local linear smoothers.}
\usage{
covariance_ll(
  data,
  U = seq(0, 1, length.out = 101),
  t0_list = seq(0.1, 0.9, by = 0.1),
  centered = FALSE
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{Vector, sampling points at which estimate the curves.}

\item{t0_list}{Vector, the sampling points at which we estimate the
parameters.}

\item{centered}{Boolean (default=FALSE), are the data centered?}
}
\value{
A list of with three entries:
\itemize{
\item \strong{$parameter} The estimated parameters.
\item \strong{$bandwidth} The estimated bandwidths.
\item \strong{$cov} The estimated covariance.
}
}
\description{
This function performs the estimation of the covariance of a set of curves
using local linear smoothers where the bandwidth is estimated using the
methodology from Golovkine et al. (2021).
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
