% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{covariance_lll}
\alias{covariance_lll}
\title{Perform an estimation of the covariance with local linear smoothers.}
\usage{
covariance_lll(data, U, b = 0.1)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{Vector, sampling points at which estimate the covariance.}

\item{b}{Numeric (default=0.1), the bandwidth to use.}
}
\value{
A matrix representing the covariance surface.
}
\description{
This function performs the estimation of the covariance of a set of curves
using local linear smoothers where the bandwidth is given.
}
\references{
Zhang X. and Wang J.-L. (2016) - From sparse to dense functional
data and beyond, The Annals of Statistics
}
