% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bandwidth.R
\name{estimate_bandwidth}
\alias{estimate_bandwidth}
\title{Perform an estimation of the bandwidth for the estimation of the mean}
\usage{
estimate_bandwidth(
  data,
  t0,
  H0 = 0.5,
  L0 = 1,
  sigma = 0,
  variance = 0,
  grid = lseq(0.001, 0.1, length.out = 101),
  nb_obs_minimal = 2,
  type_k = 2
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{t0}{Numeric, the sampling point at which we estimate the bandwidth.}

\item{H0}{Numeric, an estimation of \eqn{H_0}.}

\item{L0}{Numeric, an estimation of \eqn{L_0}.}

\item{sigma}{Numeric, an estimation of \eqn{\sigma}.}

\item{variance}{Numeric, an estimation of the variance of the process.}

\item{grid}{Vector, default=lseq(0.001, 0.1, length.out = 101). A grid of
bandwidths.}

\item{nb_obs_minimal}{Integer, default=2. Minimum number of points in the
neighborhood to keep the curve in the estimation.}

\item{type_k}{Integer, default=2. Used kernel.
\itemize{
\item \strong{1} Uniform kernel
\item \strong{2} Epanechnikov kernel
\item \strong{3} Biweight kernel
}}
}
\value{
Numeric, an estimation of the bandwidth.
}
\description{
This function performs an estimation of the bandwidth for a univariate kernel
regression estimator defined over continuous data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate bandwidth}
