% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bandwidth.R
\name{estimate_bandwidth_mean}
\alias{estimate_bandwidth_mean}
\title{Perform an estimation of the bandwidth for the estimation of the mean.}
\usage{
estimate_bandwidth_mean(
  curves,
  params,
  sigma = 0,
  grid_bandwidth = lseq(0.001, 0.1, length.out = 101),
  n_obs_min = 2,
  kernel_name = "epanechnikov"
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} Sampling points
\item \strong{$x} Observed points.
}}

\item{params}{List, estimation of the different parameters:
\itemize{
\item \strong{$point} Time point where the smoothing has been done.
\item \strong{$H} Estimated regularity.
\item \strong{$L} Estimated constant.
\item \strong{$var} Estimated variance.
}}

\item{sigma}{Numeric, estimation of the std of the noise \eqn{\sigma}.}

\item{grid_bandwidth}{Vector (default = lseq(0.001, 0.1, length.out = 101)),
grid of bandwidths.}

\item{n_obs_min}{Integer (default = 2), minimum number of points in the
neighborhood to keep the curve in the estimation.}

\item{kernel_name}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}
}
\value{
Numeric, estimation of the bandwidth.
}
\description{
This function performs an estimation of the bandwidth for a univariate kernel
regression estimator defined over continuous data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\seealso{
Other estimate bandwidth: 
\code{\link{estimate_bandwidth_covariance}()}
}
\concept{estimate bandwidth}
