% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bandwidth.R
\name{estimate_bandwidths}
\alias{estimate_bandwidths}
\title{Perform an estimation of the bandwidth for the estimation of the mean.}
\usage{
estimate_bandwidths(
  data,
  t0_list = 0.5,
  grid = NULL,
  nb_obs_minimal = 2,
  type_k = 2
)
}
\arguments{
\item{data}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{t0_list}{Vector, the sampling points at which we estimate the
parameters.}

\item{grid}{Vector, default=NULL. A grid of bandwidths.}

\item{nb_obs_minimal}{Integer, default=2. Minimum number of points in the
neighborhood to keep the curve in the estimation.}

\item{type_k}{Integer, default=2. Used kernel.
\itemize{
\item \strong{1} Uniform kernel
\item \strong{2} Epanechnikov kernel
\item \strong{3} Biweight kernel
}}
}
\value{
List, with elements:
\itemize{
\item \strong{sigma} An estimation of the standard deviation of the noise
\item \strong{variance} An estimation of the variance of the process
\item \strong{H0} An estimation of \eqn{H_0}
\item \strong{L0} An estimation of \eqn{L_0}
\item \strong{b} An estimation of the bandwidth
}
}
\description{
This function performs an estimation of the bandwidth to be used in the
Nadaraya-Watson estimator.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
