% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_constant}
\alias{estimate_constant}
\title{Perform the estimation of the constant \eqn{L_0}.}
\usage{
estimate_constant(curves_smooth, regularity)
}
\arguments{
\item{curves_smooth}{List, resulting from the \code{presmoothing} function.}

\item{regularity}{Numeric, estimation of the regularity of the curves and
resulting from the the \code{estimate_regularity} function.}
}
\value{
Numeric, an estimation of \eqn{L_0} at \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{L_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate \eqn{L_0}}
