% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_crossvar}
\alias{estimate_crossvar}
\title{Perform the estimation of the variance \eqn{Var(X_{s}X_{t)}}.}
\usage{
estimate_crossvar(curves_smooth_s, curves_smooth_t)
}
\arguments{
\item{curves_smooth_s}{List, smoothing of the curves at point \eqn{s} and
resulting from the \code{presmoothing} function.}

\item{curves_smooth_t}{List, smoothing of the curves at point \eqn{t} and
resulting from the \code{presmoothing} function.}
}
\value{
Numeric, estimation of \eqn{Var(X_{s}X_{t)}}.
}
\description{
This function performs an estimation of the variance \eqn{Var(X_{s}X_{t)}}
used for the estimation of the bandwidth for a univariate kernel regression
estimator defined over continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate variance}
