% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_curve.R
\name{estimate_curve}
\alias{estimate_curve}
\title{Perform the smoothing of an individual curve.}
\usage{
estimate_curve(
  curve,
  U,
  b,
  t0_list = NULL,
  kernel = "epanechnikov",
  n_obs_min = 1
)
}
\arguments{
\item{curve}{A list, with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{Vector, sampling points at which estimate the curve.}

\item{b}{Vector, estimation of the bandwidth. If one is provided, we use a
unique bandwidth for the curve. However, if a vector is given, the bandwidth
changes depending on the sampling points.}

\item{t0_list}{Vector, times at which the bandwidths have been
estimated. Only used if the parameter \code{b} is a vector.}

\item{kernel}{Character string, the kernel used for the estimation:
\itemize{
\item epanechnikov (default)
\item uniform
\item beta
\item mBeta
}}

\item{n_obs_min}{Integer, minimum number of observation for the smoothing.}
}
\value{
List, with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The estimated points.
}
}
\description{
This function performs the smoothing of a curve using the Nadaraya-Watson
estimator given a particular kernel.
}
