% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_density}
\alias{estimate_density}
\title{Perform an estimation of the minimum of the sampling points density.}
\usage{
estimate_density(data)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}
}
\value{
An estimation of the minimum of the density of the sampling points.
}
\description{
This function performs an estimation of the minimum of the density of the
noise.
}
