% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_H0}
\alias{estimate_H0}
\title{Perform an estimation of \eqn{H_0}.}
\usage{
estimate_H0(data)
}
\arguments{
\item{data}{List of array, resulting from the pre-smoothing function.}
}
\value{
Vector, an estimation of \eqn{H_0} at each \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{H_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate \eqn{H_0}}
