% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_L0}
\alias{estimate_L0}
\title{Perform the estimation of \eqn{L_0}.}
\usage{
estimate_L0(data, H0_list, M)
}
\arguments{
\item{data}{List of array, resulting from the pre-smoothing function.}

\item{H0_list}{Vector, resulting from estimate_H0 function.}

\item{M}{Numeric, mean number of sampling points per curve.}
}
\value{
Vector, estimation of \eqn{L_0}.
}
\description{
This function performs an estimation of \eqn{L_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate \eqn{L_0}}
