% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_moment}
\alias{estimate_moment}
\title{Perform the estimation of the moments.}
\usage{
estimate_moment(data, order = 1)
}
\arguments{
\item{data}{List of array, resulting from the pre-smoothing function.}

\item{order}{Numeric (default=1), the moment to estimate.}
}
\value{
Vector, estimation of the moments.
}
\description{
This function performs an estimation of the moments used for the estimation
of the bandwidth for a univariate kernel regression estimator defined over
continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate moment}
