% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters_covariance}
\alias{estimate_parameters_covariance}
\title{Perform a recursive estimation of the parameters over a grid of points for
the estimation of the covariance.}
\usage{
estimate_parameters_covariance(
  curves,
  grid = c(0.25, 0.5, 0.75),
  delta_f = NULL,
  kernel_name = "epanechnikov",
  beta = 1
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{grid}{Vector (default = c(0.25, 0.5, 0.75)), sampling points at which
the data is pre-smoothed.}

\item{delta_f}{Function (default = NULL), function to determine the delta.}

\item{kernel_name}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}

\item{beta}{Numeric (default = 1), pre-specified regularity of the curves to
start the recursion. The default value is 1, which correspond to at least one
time differentiable curves.}

\item{compute_crossvar}{Boolean (default = FALSE), should
\eqn{Var(X_{s}X_{t)}} be computed for covariance estimation?}
}
\value{
Dataframe, with columns:
\itemize{
\item \strong{$point} Time point where the smoothing has been done.
\item \strong{$curves} Smoothed curves.
\item \strong{$H} Estimated regularity.
\item \strong{$L} Estimated constant.
\item \strong{$var} Estimated variance.
\item \strong{$mom} Estimated \eqn{E(X^{2}_{t_0})}
\item \strong{$var_st} \eqn{Var(X_{s}X_{t)}}
}
}
\description{
This function performs a recursive estimation of the different parameters
used for the estimation of the covariance estimation of functional data. The
recursion is made by small step onto the estimation of the regularity of the #' curves. The pre-smoothing of the data is done using a Nadaraya-Watson
estimator and the used bandwidth modified using each new estimation of the
regularity.
}
\references{
S. Golovkine, N. Klutchnikoff and V. Patilea (2021) - Adaptive
optimal estimation of irregular mean and covariance functions.
}
