% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters_recursion}
\alias{estimate_parameters_recursion}
\title{Perform a recursive estimation of the parameters.}
\usage{
estimate_parameters_recursion(
  curves,
  point = 0.5,
  delta_f = NULL,
  kernel = "epanechnikov",
  beta = 1
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{point}{Numeric (default = 0.5), sampling point at which the data is
pre-smoothed.}

\item{delta_f}{Function (default = NULL), function to determine the delta.}

\item{kernel}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}

\item{beta}{Numeric (default = 1), pre-specified regularity of the curves to
start the recursion. The default value is 1, which correspond to at least one
time differentiable curves.}
}
\value{
List, with four entries:
\itemize{
\item \strong{$point} Time point where the smoothing has been done.
\item \strong{$H} Estimated regularity.
\item \strong{$L} Estimated constant.
\item \strong{$var} Estimated variance.
}
}
\description{
This function performs a recursive estimation of the different parameters
used for the estimation of the mean and covariance estimation of functional
data. The recursion is made by small step onto the estimation of the
regularity of the curves. The pre-smoothing of the data is done using a
Nadaraya-Watson estimator and the used bandwidth modified using each new
estimation of the regularity.
}
\references{
S. Golovkine, N. Klutchnikoff and V. Patilea (2021) - Adaptive
optimal estimation of irregular mean and covariance functions.
}
