% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_regularity}
\alias{estimate_regularity}
\title{Perform an estimation of the regularity \eqn{H_0}.}
\usage{
estimate_regularity(curves_smooth)
}
\arguments{
\item{curves_smooth}{List, resulting from the \code{presmoothing} function.}
}
\value{
Numeric, an estimation of \eqn{H_0} at \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{H_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate \eqn{H_0}}
