% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Perform an estimation of the standard deviation of the noise.}
\usage{
estimate_sigma(data, t0_list, k0_list)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{t0_list}{Vector, the sampling points at which we estimate \eqn{H_0}. We
will consider the \eqn{8k0 - 7} nearest points of \eqn{t_0} for the
estimation of \eqn{H_0} when \eqn{\sigma} is unknown.}

\item{k0_list}{Vector, the number of neighbors of \eqn{t_0} to consider.
Should be set as \eqn{k0 = M\exp(-(\log(\log(M))**2))}. We can set a
different \eqn{k_0}, but in order to use the same for each \eqn{t_0}, just
put a unique numeric.}
}
\value{
A list, an estimation of sigma at different \eqn{t_0}.
}
\description{
This function performs an estimation of the standard deviation of the noise
in the curves. The following formula is used:
\deqn{\hat{\sigma^2} = \frac{1}{N}\sum_{n = 1}^{N} 
      \frac{1}{2(M_n - 1)}\sum_{l = 2}^{M_n}(Y_{n, (l)} - Y_{n, (l-1)})^2}
}
\references{
S. Golovkine, N. Klutchnikoff, V. Patilea (2020) - Learning the
smoothness of noisy curves with application to online curve estimation.
}
