% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_var}
\alias{estimate_var}
\title{Perform an estimation of \eqn{Var(X_{t_0)}}.}
\usage{
estimate_var(data)
}
\arguments{
\item{data}{List of array, resulting from the pre-smoothing function.}
}
\value{
Vector, estimation of the variance at each \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{Var(X_{t_0})} used for the
estimation of the bandwidth for the mean and the covariance by a univariate
kernel regression estimator.
}
\references{
S. Golovkine, N. Klutchnikoff and V. Patilea (2021) - Adaptive
optimal estimation of irregular mean and covariance functions.
}
