% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{irregFunData2list}
\alias{irregFunData2list}
\title{Convert \code{funData::irregFunData} objects into the right format.}
\usage{
irregFunData2list(data, norm = TRUE)
}
\arguments{
\item{data}{An object of the class \code{funData::irregFunData}.}

\item{norm}{Boolean, if TRUE, the sampling points are normalized on
\eqn{[0, 1]}}
}
\value{
A list, where each element represents a curve. Each curve is defined
as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item
\strong{$x} The observed points
}
}
\description{
Convert \code{funData::irregFunData} objects into the right format.
}
\references{
C. Happ-Kurz (2020) Object-Oriented Software for Functional Data.
Journal of Statistical Software, 93(5): 1-38 .
}
