% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernel}
\alias{kernel}
\title{Compute the values for some kernels.}
\usage{
kernel(u, type = 1)
}
\arguments{
\item{u}{Vector, points to estimate the kernel.}

\item{type}{Integer, used kernel.
\itemize{
\item \strong{1} Uniform kernel
\item \strong{2} Epanechnikov kernel
\item \strong{3} Biweight kernel
}}
}
\value{
Vector, the evaluated kernel.
}
\description{
Compute the values for some kernels.
}
