% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernel_f}
\alias{kernel_f}
\title{Compute the values for some kernels.}
\usage{
kernel_f(u, type = "epanechnikov")
}
\arguments{
\item{u}{Vector, points to estimate the kernel.}

\item{type}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}
}
\value{
Vector, the evaluated kernel.
}
\description{
Compute the values for some kernels.
}
