% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lseq}
\alias{lseq}
\title{Logarithmic sequence generation.}
\usage{
lseq(from = 1, to = 100, length.out = 51)
}
\arguments{
\item{from}{Numeric, the starting values of the sequence.}

\item{to}{Numeric, the end value of the sequence.}

\item{length.out}{Numeric, desired length of the sequence.}
}
\value{
Vector, the logarithmic sequence.
}
\description{
Logarithmic sequence generation.
}
