% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_mean.R
\name{mean_lll}
\alias{mean_lll}
\title{Perform an estimation of the mean with local linear smoothers.}
\usage{
mean_lll(data, U)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{Vector, sampling points at which estimate the mean.}
}
\value{
A vector representing the mean curve.
}
\description{
This function performs the estimation of the mean of a set of curves using
local linear smoothers where the bandwidth is estimated using the
methodology from Zhang et Wang (2016).
}
\references{
Zhang X. and Wang J.-L. (2016) - From sparse to dense functional
data and beyond, The Annals of Statistics
}
