% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neighbors}
\alias{neighbors}
\title{Test whether the observation points are in the neighborhood.}
\usage{
neighbors(t, t0, h, k0)
}
\arguments{
\item{t}{Vector, the sampling points.}

\item{t0}{Numeric, the estimation point.}

\item{h}{Numeric, the bandwidth.}

\item{k0}{Numeric, the number of points to be in the neighborhood.}
}
\value{
Boolean, true if there are enough points in the neighborhood.
}
\description{
Test if there are more than \eqn{k_0} points, defined by \eqn{t}, in the
neighborhood of \eqn{t_0}, defined by the bandwidth \eqn{h}.
}
