% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{presmoothing}
\alias{presmoothing}
\title{Perform a pre-smoothing of the data.}
\usage{
presmoothing(data, t0_list = 0.5, gamma = 0.5, order = 1, drv = 0)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{t0_list}{Vector, the sampling point at which we pre-smooth the data.}

\item{gamma}{Numeric, default=0.5. Constant \eqn{\gamma} used in the theorem
1 in the paper. Should be between 0 and 1.}

\item{order}{Integer, default=0. Regularity of the input data.}

\item{drv}{Integer, default=0. Order of derivative to be estimated.}
}
\value{
List of array. Contains the smoothed data.
}
\description{
This function performs a pre-smoothing of the data by local linear smoother.
We use a Gaussian kernel and a naive bandwidth.
}
\references{
S. Golovkine, N. Klutchnikoff and V. Patilea (2021) - Adaptive
optimal estimation of irregular mean and covariance functions.
}
