% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_curve.R
\name{smooth_curves}
\alias{smooth_curves}
\title{Perform a non-parametric smoothing of a set of curves for mean estimation.}
\usage{
smooth_curves(
  data,
  U = NULL,
  t0_list = 0.5,
  grid = NULL,
  nb_obs_minimal = 2,
  kernel = "epanechnikov"
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{Vector, default=NULL. Sampling points at which estimate the curves.
If NULL, the sampling points for the estimation are the same than the
observed ones.}

\item{t0_list}{Vector, the sampling points at which we estimate the
parameters.}

\item{grid}{Vector, default=NULL. A grid of bandwidths.}

\item{nb_obs_minimal}{Integer, minimum number of observation for the smoothing.}

\item{kernel}{Character string, the kernel used for the estimation:
\itemize{
\item epanechnikov (default)
\item uniform
\item biweight
}}
}
\value{
A list, which contains two elements. The first one is a list which
contains the estimated parameters:
\itemize{
\item \strong{sigma} An estimation of the standard deviation of the noise
\item \strong{variance} An estimation of the variance of the process
\item \strong{H0} An estimation of \eqn{H_0}
\item \strong{L0} An estimation of \eqn{L_0}
\item \strong{b} An estimation of the bandwidth
}
The second one is another list which contains the estimation of the curves:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The estimated points.
}
}
\description{
This function performs a non-parametric smoothing of a set of curves using
the Nadaraya-Watson estimator.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
