% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_curve.R
\name{smooth_curves_mean}
\alias{smooth_curves_mean}
\title{Perform a non-parametric smoothing of a set of curves for mean estimation.}
\usage{
smooth_curves_mean(
  curves,
  grid = NULL,
  grid_param = c(0.25, 0.5, 0.75),
  grid_bandwidth = NULL,
  delta_f = NULL,
  kernel_name = "epanechnikov",
  n_obs_min = 2
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed points.
}}

\item{grid}{Vector (default = NULL), sampling points at which estimate the
curves. If NULL, the sampling points for the estimation are the same than the
observed ones.}

\item{grid_param}{Vector (default = c(0.25, 0.5, 0.75)), sampling points at
which we estimate the parameters.}

\item{grid_bandwidth}{Vector (default = NULL), grid of bandwidths.}

\item{delta_f}{Function (default = NULL), function to determine the delta.}

\item{kernel_name}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}

\item{n_obs_min}{Integer (default = 2), minimum number of observation for
the smoothing.}
}
\value{
A list, which contains two elements. The first one is a list which
contains the estimated parameters:
\itemize{
\item \strong{sigma} Estimation of the standard deviation of the noise.
\item \strong{variance} Estimation of the variance of the process.
\item \strong{H0} Estimation of \eqn{H_0}.
\item \strong{L0} Estimation of \eqn{L_0}.
\item \strong{b} Estimation of the bandwidth.
}
The second one is another list which contains the estimation of the curves:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Estimated points.
}
}
\description{
This function performs a non-parametric smoothing of a set of curves using
the Nadaraya-Watson estimator.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
