% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{variance}
\alias{variance}
\title{Perform the estimation of \eqn{Var(X_{s}X_{t)}}.}
\usage{
variance(data)
}
\arguments{
\item{data}{List of array, resulting from the pre-smoothing function.}
}
\value{
Vector, estimation of \eqn{Var(X_{s}X_{t)}}.
}
\description{
This function performs an estimation of  \eqn{Var(X_{s}X_{t)}} used for the
estimation of the bandwidth for a univariate kernel regression estimator
defined over continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate variance}
