% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{covariance_ll}
\alias{covariance_ll}
\title{Perform an estimation of the covariance with local linear smoothers.}
\usage{
covariance_ll(
  curves,
  grid = seq(0, 1, length.out = 101),
  grid_param = seq(0.1, 0.9, by = 0.1),
  grid_bandwidth = NULL,
  centered = FALSE,
  delta_f = NULL,
  n_obs_min = 2,
  kernel_name = "epanechnikov"
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} Sampling points
\item \strong{$x} Observed points.
}}

\item{grid}{Vector (default = seq(0, 1, length.out = 101)), sampling points
at which estimate the curves.}

\item{grid_param}{Vector (default = seq(0.1, 0.9, by = 0.1)), the sampling
points at which we estimate the parameters.}

\item{grid_bandwidth}{Vector (default = NULL), grid of bandwidths. -> MEAN}

\item{delta_f}{Function (default = NULL), function to determine the delta. -> MEAN}

\item{n_obs_min}{Integer (default = 2), minimum number of observation for the
smoothing. -> MEAN}

\item{kernel_name}{String (default = 'epanechnikov'), the kernel used for the
estimation: -> MEAN
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}

\item{center}{Boolean (default = TRUE), center the data?}
}
\value{
List of with three entries:
\itemize{
\item \strong{$parameter} The estimated parameters.
\item \strong{$bandwidth} The estimated bandwidths.
\item \strong{$cov} The estimated covariance.
}
}
\description{
This function performs the estimation of the covariance of a set of curves
using local linear smoothers where the bandwidth is estimated using the
methodology from Golovkine et al. (2021).
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
