% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_covariance.R
\name{covariance_lll}
\alias{covariance_lll}
\title{Perform an estimation of the covariance with local linear smoothers.}
\usage{
covariance_lll(curves, grid, bandwidth = 0.1)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} Sampling points
\item \strong{$x} Observed points.
}}

\item{grid}{Vector, sampling points at which estimate the covariance.}

\item{bandwidth}{Numeric (default = 0.1), the bandwidth to use.}
}
\value{
Matrix representing the covariance surface.
}
\description{
This function performs the estimation of the covariance of a set of curves
using local linear smoothers where the bandwidth is given.
}
\references{
Zhang X. and Wang J.-L. (2016) - From sparse to dense functional
data and beyond, The Annals of Statistics
}
