% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_curve.R
\name{estimate_curve}
\alias{estimate_curve}
\title{Perform the smoothing of an individual curve.}
\usage{
estimate_curve(
  curve,
  grid,
  bandwidth,
  bandwidth_times = NULL,
  kernel_name = "epanechnikov",
  n_obs_min = 1
)
}
\arguments{
\item{curve}{List, with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Observed points.
}}

\item{grid}{Vector, sampling points at which the curve is estimated.}

\item{bandwidth}{Vector, estimation of the bandwidth. If a unique element is
provided, we use a  unique bandwidth for the curve. However, if a vector is
given, the bandwidth changes depending on the sampling points.}

\item{bandwidth_times}{Vector (default = NULL), times at which the bandwidths
have been estimated. Only used if the parameter \code{bandwidth} is a vector.}

\item{kernel_name}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}

\item{n_obs_min}{Integer (default = 1), minimum number of observation for
the smoothing.}
}
\value{
List, with two entries:
\itemize{
\item \strong{$t} Sampling points.
\item \strong{$x} Estimated points.
}
}
\description{
This function performs the smoothing of a curve using the Nadaraya-Watson
estimator given a particular kernel.
}
