% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_moment}
\alias{estimate_moment}
\title{Perform the estimation of the moments.}
\usage{
estimate_moment(curves_smooth, order = 1)
}
\arguments{
\item{curves_smooth}{List, resulting from the \code{presmoothing} function.}

\item{order}{Numeric (default = 1), the moment to estimate.}
}
\value{
Numeric, an estimation of the moments \eqn{E(X^{\alpha}_{t_0})}
}
\description{
This function performs an estimation of the moments \eqn{E(X^{\alpha}_{t_0})}
used for the estimation of the bandwidth for a univariate kernel regression
estimator defined over continuous domains data.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\concept{estimate moment}
