% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Perform an estimation of the standard deviation of the noise.}
\usage{
estimate_sigma(data, delta = 0.1)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{delta}{Numeric (default = 0.1), neighborhood for the estimation.}
}
\value{
A list, an estimation of sigma at different \eqn{t_0}.
}
\description{
This function performs an estimation of the standard deviation of the noise
in the curves. The following formula is used:
\deqn{\hat{\sigma^2} = \frac{1}{N}\sum_{n = 1}^{N} 
      \frac{1}{2(M_n - 1)}\sum_{l = 2}^{M_n}(Y_{n, (l)} - Y_{n, (l-1)})^2}
}
\references{
S. Golovkine, N. Klutchnikoff, V. Patilea (2020) - Learning the
smoothness of noisy curves with application to online curve estimation.
}
