% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{presmoothing}
\alias{presmoothing}
\title{Perform a pre-smoothing of the data.}
\usage{
presmoothing(
  curves,
  point = 0.5,
  delta_f = NULL,
  kernel = "epanechnikov",
  beta = 1,
  bandwidth_naive = 0
)
}
\arguments{
\item{curves}{List, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} Sampling points
\item \strong{$x} Observed points
}}

\item{point}{Numeric (default = 0.5), sampling point at which the data is
pre-smoothed.}

\item{delta_f}{Function (default = NULL), function to determine the delta.}

\item{kernel}{String (default = 'epanechnikov'), the kernel used for the
estimation:
\itemize{
\item epanechnikov
\item uniform
\item biweight
}}

\item{beta}{Numeric (default = 1), pre-specified regularity of the curves.
The default value is 1, which correspond to at least one time
differentiable curves.}

\item{bandwidth_naive}{Numeric (default = 0), bandwidth to use for the
presmoothing. If set to 0, the bandwidth will be defined as
\deqn{\frac{\delta}{m}^{1 / (2\beta + 1)}}
where
\itemize{
\item \eqn{m} is the mean number of sampling points per curve.
\item \eqn{\delta} is the length of the interval where the smoothing is
done.
\item \eqn{\beta} represents the regularity of the curves.
}}
}
\value{
List, with two entries:
\itemize{
\item \strong{$grid} Grid on which the smoothing has been done.
\item \strong{$x_smooth} The smoothed data.
}
}
\description{
This function performs a pre-smoothing of the data using a Nadaraya-Watson
estimator. We use an Epanechnikov kernel and a naive bandwidth.
}
\references{
S. Golovkine, N. Klutchnikoff and V. Patilea (2021) - Adaptive
optimal estimation of irregular mean and covariance functions.
}
